+++ src/csl-engine.js
@@ -1,5 +1,6 @@
 const CSL = require('citeproc');
 const fs = require('fs');
+const dict = require('./nls');
 
 module.exports = class CSLEngine{
     constructor(extensionPath){
@@ -8,6 +9,8 @@ module.exports = class CSLEngine{
         this.citeprocSys = new CiteprocSys(this);
         this.citables = null;
         this.citablesIds = null;
+        this.lang = null;
+        this.forcedLang = null;
     }
     updateCitables(citables){
         this.citables = citables;
@@ -20,7 +23,7 @@ module.exports = class CSLEngine{
         return errors
     }
     buildProcessor(style){
-        this.proc = new CSL.Engine(this.citeprocSys, style);
+        this.proc = new CSL.Engine(this.citeprocSys, style, this.forcedLang, this.forcedLang);
         this.proc.updateItems(this.citablesIds);
     }
     buildPreviewContent(style){
@@ -54,6 +57,7 @@ class CiteprocSys{
         this.engine = engine;
     }
     retrieveLocale(lang){
+        this.engine.lang = lang;
         let path = this.engine.extensionPath + '/src/resources/locales/';
         let file = 'locales-' + lang + '.xml';
         return fs.readFileSync(path + file).toString('utf-8');
@@ -64,11 +68,11 @@ class CiteprocSys{
 }
 
 function formatHtmlPreview(individualCitations, uniqueCitation, bibliography){
-    let html = '<h3>Individual citations</h3><hr>';
+    let html = `<h3>${dict['common.indCitations']}</h3><hr>`;
     html += '<p>' + individualCitations.join('<br>') + '</p>';
-    html += '<h3>Unique citation</h3><hr>';
+    html += `<h3>${dict['common.uniqCitation']}</h3><hr>`;
     html += '<p>' + uniqueCitation[1][0][1] + '</p>';
-    html += '<h3>Bibliography</h3><hr>';
+    html += `<h3>${dict['common.Bibliography']}</h3><hr>`;
     html += bibliography;
     return html
 }