+++ src/extension.js
@@ -1,6 +1,7 @@
 const vscode = require('vscode');
 const PreviewManager = require('./previews-manager');
 const ExtensionCommander = require('./extension-commander');
+const LocaleBar = require('./locale-bar');
 
 /**
  * @param {vscode.ExtensionContext} context
@@ -8,12 +9,14 @@ const ExtensionCommander = require('./extension-commander');
 
 function activate(context) {
 	const register = vscode.commands.registerCommand;
-	const manager = new PreviewManager(context.extensionPath);
+	const localeBar = new LocaleBar(context.extensionPath);
+	const manager = new PreviewManager(context.extensionPath, localeBar);
 	const commander = new ExtensionCommander(manager);
 	context.subscriptions.push(
 		register('cslPreview.showCslPreview', () => commander.showCslPreview()),
 		register('cslPreview.refreshCslPreview', () => commander.refreshPreview()),
-		register('cslPreview.showPreviewSource', () => commander.showSource())
+		register('cslPreview.showPreviewSource', () => commander.showSource()),
+		register('cslPreview.chooseLocale', () => commander.chooseLocale())
 	);
 }
 exports.activate = activate;