+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Change Log
 
+## 0.2.0
+
+New functionality additions:
+
+* Option to choose a custom citables source JSON file
+* Support to select the locale displayed in preview
+* Added a status bar item showing the current locale displayed in preview
+* Support to VSCode locales (currently supporting en-US and pt-BR)
+* Extension configurations providing options do define default locale and citables source file
+
+Bug fix:
+
+* Command to refresh preview not more available after preview dispose
+
 ## 0.1.0
 
 Initial release of CSL Preview for Visual Studio Code